requirements.txt
aiohttp~=3.8.1
cx_Oracle~=8.3.0
cx_Oracle_async~=0.3.3
PyYAML~=6.0
aiosqlite~=0.17.0
SQLAlchemy~=1.4.36
